+++ src/Mod/TechDraw/Gui/QGVPage.h
@@ -122,6 +122,8 @@ public:
 
     void pseudoContextEvent();
 
+    QPixmap prepareCursorPixmap(const char* iconName, QPoint& hotspot);
+
     void centerOnPage();
 
     TechDraw::DrawView* getBalloonParent() { return m_balloonParent; }
@@ -150,8 +152,6 @@ protected:
     void contextMenuEvent(QContextMenuEvent* event) override;
 
     QColor getBackgroundColor();
-
-    QPixmap prepareCursorPixmap(const char* iconName, QPoint& hotspot);
 
     void drawForeground(QPainter* painter, const QRectF& rect) override;
