+++ src/Mod/TechDraw/Gui/TechDrawHandler.cpp
@@ -37,6 +37,7 @@
 #include <Gui/MainWindow.h>
 #include <Gui/View3DInventor.h>
 #include <Gui/View3DInventorViewer.h>
+#include <Gui/InputHint.h>
 
 #include "MDIViewPage.h"
 #include "QGVPage.h"
@@ -82,6 +83,11 @@ void TechDrawHandler::deactivate()
         }
         mdi->enableContextualMenu(true);
     });
+}
+
+std::list<Gui::InputHint> TechDrawHandler::getToolHints() const
+{
+    return {};
 }
 
 void TechDrawHandler::keyReleaseEvent(QKeyEvent* event)
